+++ sbnz-integracija-projekta/model/src/main/java/com/ftn/sbnz/model/models/FaultProblems/InvertorProblem.java
@@ -5,7 +5,7 @@ public class InvertorProblem  extends PowertrainProblems{
     private double resistance;
 
 
-    public InvertorProblem(double resistance, long breakdownId) {
+    public InvertorProblem(double resistance, int breakdownId) {
         this(breakdownId );
         this.resistance = resistance;
 
@@ -20,7 +20,7 @@ public class InvertorProblem  extends PowertrainProblems{
     }
 
 
-    public InvertorProblem(long breakdownId) {
+    public InvertorProblem(int breakdownId) {
         super(breakdownId);
     }
