+++ sbnz-integracija-projekta/model/src/main/java/com/ftn/sbnz/model/models/Util.java
@@ -29,21 +29,21 @@ public class Util {
       return vals.get(RANDOM.nextInt(vals.size()));
   }
 
-  public static PowertrainProblems checkPowertrain(long breakdownId){
+  public static PowertrainProblems checkPowertrain(int breakdownId){
     if(RANDOM.nextBoolean())
         return new MotorProblem(RANDOM.nextInt(100), breakdownId);
     else
         return new InvertorProblem(RANDOM.nextDouble()*4000.0, breakdownId);
   }
 
-public static FaultProblem checkAuxiliary(long breakdownId) {
+public static FaultProblem checkAuxiliary(int breakdownId) {
      if(RANDOM.nextBoolean())
         return new HVACProblem(randomHVACProblem(), breakdownId);
     else
         return new InfotainmentProblem(breakdownId);
 }
 
-public static FaultProblem checkRegenerativeBreaking(long breakdownId) {
+public static FaultProblem checkRegenerativeBreaking(int breakdownId) {
      if(RANDOM.nextBoolean())
         return new WheelSpeedProblem(RANDOM.nextDouble()*100.0, breakdownId);
     else
